So there are still a ton of issues - I merged main into buildPreferUpHeuristic branch, and I manually have been going in and updating classes. I deleted module Game2048Collect, and moved classes into Game2048Base. I renamed PlayGame class ot PlayGameManual, and deleted the this branch's previous PlayGameManual. There are a lot of conflicts in the Intelligence Module - such that classes are trying ot access Grid2048's grid, which was a huge part of the develop branch pr. I will be updating that next.
