--- Game2048Base/src/main/java/utility/PosnUtility.java
@@ -1,25 +1,0 @@
-package utility;
-
-import javalib.worldimages.Posn;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Random;
-
-public class PosnUtility {
-
-    public static ArrayList<Posn> createRandomPosns() {
-        ArrayList<Posn> result;
-        Random r = new Random();
-
-        Posn p1 = new Posn(r.nextInt(4), r.nextInt(4));
-        Posn p2 = new Posn(r.nextInt(4), r.nextInt(4));
-
-        if (p1.equals(p2)) {
-            result = createRandomPosns();
-        } else {
-            result = new ArrayList<>(Arrays.asList(p1, p2));
-        }
-
-        return result;
-    }
-}
