+++ Game2048Player/src/main/java/manual/PlayGameManual.java
@@ -1,36 +1,39 @@
-package manual;
+package playgame;
 
-import javalib.funworld.*;
-import javalib.worldimages.*;
-import game2048.Game2048;
-import game2048.Scoreboard;
-import models.board2048.Board2048;
+import javalib.funworld.World;
+import javalib.funworld.WorldScene;
+import javalib.worldimages.TextImage;
+import javalib.worldimages.WorldEnd;
+import models.game2048.Game2048;
+import models.game2048.KeyEventHandler;
+import models.game2048.Scoreboard;
+import models.grid2048.Grid2048;
 import models.square.Square;
-import java.awt.*;
+import java.awt.Color;
 
-public class PlayGameManual extends World {
-    public Game2048 game2048;
+public class PlayGame extends World {
+    private final Game2048 game2048;
 
-    public PlayGameManual() {}
+    public PlayGame(Game2048 game2048) {
+        this.game2048 = game2048;
+    }
 
     public static void main (String[] args) {
-        PlayGameManual g = new PlayGameManual();
-        Game2048 game2048 = new Game2048(new Board2048().initializeStartingGrid(), new Scoreboard(0));
-        g.setGame2048(game2048);
+        Game2048 newGame = new Game2048(new Grid2048(), new Scoreboard(0));
+        PlayGame g = new PlayGame(newGame);
         g.bigBang(Square.SIDE_LENGTH * 6,Square.SIDE_LENGTH * 6,1);
     }
 
     @Override
     public WorldScene makeScene() {
         WorldScene s = new WorldScene(Square.SIDE_LENGTH * 6, Square.SIDE_LENGTH * 6);
-        return s.placeImageXY(this.getGame2048().getBoard2048().drawGrid(), Square.SIDE_LENGTH * 3, Square.SIDE_LENGTH * 3)
+        return s.placeImageXY(this.getGame2048().getGrid2048().drawGrid(), Square.SIDE_LENGTH * 3, Square.SIDE_LENGTH * 3)
                 .placeImageXY(this.getGame2048().getScoreboard().drawScoreboard(), Scoreboard.WIDTH/2 + 10 , Scoreboard.HEIGHT);
     }
 
     @Override
     public WorldScene lastScene (String msg) {
-        return this.makeScene()
-                .placeImageXY(new TextImage(msg, 100, Color.BLACK), Square.SIDE_LENGTH * 3, Square.SIDE_LENGTH * 3);
+        return this.makeScene().placeImageXY(new TextImage(msg, 100, Color.BLACK), Square.SIDE_LENGTH * 3, Square.SIDE_LENGTH * 3);
     }
 
     @Override
@@ -44,55 +47,41 @@ public class PlayGameManual extends World {
 
     @Override
     public World onKeyEvent(String s) {
-
-        PlayGameManual updatedPlayGameManual = new PlayGameManual();
-        Game2048 updatedGame2048;
-
-        if (s.equals("up")) {
-            updatedGame2048 = this.getGame2048().handleUpEvent().getUpdatedGame2048();
-            updatedPlayGameManual.setGame2048(updatedGame2048);
-            return updatedPlayGameManual;
-        }
-        if (s.equals("down")) {
-            updatedGame2048 = this.getGame2048().handleDownEvent().getUpdatedGame2048();
-            updatedPlayGameManual.setGame2048(updatedGame2048);
-            return updatedPlayGameManual;
-        }
+        Grid2048 currentGrid = game2048.getGrid2048();
         if (s.equals("left")) {
-            updatedGame2048 = this.getGame2048().handleLeftEvent().getUpdatedGame2048();
-            updatedPlayGameManual.setGame2048(updatedGame2048);
-            return updatedPlayGameManual;
+            KeyEventHandler handler = currentGrid.handleLeftEvent(game2048.getScoreboard());
+            return new PlayGame(new Game2048(handler.getGrid2048(), handler.getScoreboard()));
         }
         if (s.equals("right")) {
-            updatedGame2048 = this.getGame2048().handleRightEvent().getUpdatedGame2048();
-            updatedPlayGameManual.setGame2048(updatedGame2048);
-            return updatedPlayGameManual;
-
+            KeyEventHandler handler = currentGrid.handleRightEvent(game2048.getScoreboard());
+            return new PlayGame(new Game2048(handler.getGrid2048(), handler.getScoreboard()));
+        }
+        if (s.equals("down")) {
+            KeyEventHandler handler = currentGrid.handleDownEvent(game2048.getScoreboard());
+            return new PlayGame(new Game2048(handler.getGrid2048(), handler.getScoreboard()));
+        }
+        if (s.equals("up")) {
+            KeyEventHandler handler = currentGrid.handleUpEvent(game2048.getScoreboard());
+            return new PlayGame(new Game2048(handler.getGrid2048(), handler.getScoreboard()));
         } else {
-            return this; }
-
+            return this;
+        }
     }
 
     boolean isGameOver () {
+        Grid2048 currentGrid = game2048.getGrid2048();
+        Grid2048 gridUp = currentGrid.handleUpEvent(game2048.getScoreboard()).getGrid2048();
+        Grid2048 gridDown = currentGrid.handleDownEvent(game2048.getScoreboard()).getGrid2048();
+        Grid2048 gridRight = currentGrid.handleRightEvent(game2048.getScoreboard()).getGrid2048();
+        Grid2048 gridLeft = currentGrid.handleLeftEvent(game2048.getScoreboard()).getGrid2048();
 
-        Board2048 gridUp = this.getGame2048().handleUpEvent().getUpdatedGame2048().getBoard2048();
-        Board2048 gridDown = this.getGame2048().handleDownEvent().getUpdatedGame2048().getBoard2048();
-        Board2048 gridRight = this.getGame2048().handleRightEvent().getUpdatedGame2048().getBoard2048();
-        Board2048 gridLeft = this.getGame2048().handleLeftEvent().getUpdatedGame2048().getBoard2048();
-        Board2048 currentBoard = this.getGame2048().getBoard2048();
-
-        return
-                gridUp.equals(currentBoard) &&
-                        gridDown.equals(currentBoard) &&
-                        gridRight.equals(currentBoard) &&
-                        gridLeft.equals(currentBoard);
+        return  gridUp.equals(currentGrid)
+                && gridDown.equals(currentGrid)
+                && gridRight.equals(currentGrid)
+                && gridLeft.equals(currentGrid);
     }
 
     public Game2048 getGame2048() {
         return game2048;
-    }
-
-    public void setGame2048(Game2048 game2048) {
-        this.game2048 = game2048;
     }
 }
