+++ Game2048Player/src/test/java/manual/PlayGameManualTest.java
